+++ resources/assets/js/admin/display/actions.js
@@ -6,11 +6,18 @@ Admin.Modules.register('display.actions', () => {
         $.ajax({
             type: $selectActions.find('option:selected').data('method'),
             url: $selectActions.val(),
-            data: $checkboxes.serialize()
-
+            data: $checkboxes.serialize(),
+            dataType: 'json'
         }).done(function (msg) {
-                // TODO Add success message
-            });
+            if (msg.hasOwnProperty('text')) {
+                swal({title: msg.text, text: msg.message, type: msg.type, timer: 5000})
+                setTimeout(function(){
+                    location.reload();
+                    window.location.reload();
+                }, 5000);
+
+            }
+        });
 
         return false;
     });